fixing linting and more refactoring
